backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.120.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
